chore: generate command-not-found db instead of storing it as header files in git history
From now onwards, there will be a new version of command-not-found only
when a new feature gets added to the program, or something gets changed.
Unlike earlier when new version of command-not-found was released when
the database was updated, now this won't happen. Instead just a revbump
of command-not-found will be done, so that users can get the new
commands.

On the packaging side, we just need to run term_setup_nodejs and run the
'generate-db.js' in source directory. In future a more reliant method
will be added so that other apt-packaged distributions can also use
command-not-found.